% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS.R
\name{GVS}
\alias{GVS}
\title{Check the validity of coordinates}
\usage{
GVS(occurrence_dataframe, ...)
}
\arguments{
\item{occurrence_dataframe}{A properly formatted dataframe, see \code{gvs_testfile}}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing GVS results.
}
\description{
GVS returns information on coordinate validity.
}
\examples{
\donttest{

results <- GVS(occurrence_dataframe = gvs_testfile)

}
}
