% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS_citations.R
\name{GVS_citations}
\alias{GVS_citations}
\title{Get citation information}
\usage{
GVS_citations(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing bibtex-formatted citation information
}
\description{
Returns information needed to cite the GVS
}
\note{
This function provides citation information in bibtex format that can be used with reference manager software (e.g. Paperpile, Zotero). Please do remember to cite both the sources and the GVS, as the GVS couldn't exist without these sources!
}
\examples{
{
citation_info <- GVS_citations()
}

}
