% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS_collaborators.R
\name{GVS_collaborators}
\alias{GVS_collaborators}
\title{Get collaborator information}
\usage{
GVS_collaborators(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing bibtex-formatted citation information
}
\description{
Returns information on GVS collaborators
}
\examples{
{
collaborator_info <- GVS_collaborators()
}

}
