% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS_data_dictionary.R
\name{GVS_data_dictionary}
\alias{GVS_data_dictionary}
\title{Get data dictionary}
\usage{
GVS_data_dictionary(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing bibtex-formatted citation information
}
\description{
Returns the GVS data dictionary
}
\examples{
{
data_dictionary <- GVS_data_dictionary()
}

}
