% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS_sources.R
\name{GVS_sources}
\alias{GVS_sources}
\title{Get information on sources used by the GVS}
\usage{
GVS_sources(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing information about the sources used in the current GVS version.
}
\description{
Return metadata about the current GVS sources
}
\examples{
{
sources <- GVS_sources()
}

}
