% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvs_testfile}
\alias{gvs_testfile}
\title{Example GVS data}
\format{
A data.frame with 27 observations of 2 variables:
\describe{
\item{Latitude}{Latitude, in decimal degrees}
\item{Longitude}{Longitude, in decimal degrees}
...
}
}
\source{
\url{https://biendata.org}
}
\usage{
gvs_testfile
}
\description{
A sample dataset showing the proper formatting of GVS inputs.
}
\keyword{datasets}
