% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVS_version.R
\name{GVS_version}
\alias{GVS_version}
\title{Get metadata on current GVS version}
\usage{
GVS_version(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing current GVS version number, build date, and code version.
}
\description{
Return metadata about the current GVS version
}
\examples{
{
NSR_version_metadata <- GVS_version()
}

}
